changed deezer_id to bigint, added label_id to track model, added bullmq package, added queue for label-tracks and artist albums

+++ backend/models/album.js
@@ -32,7 +32,7 @@ module.exports = (sequelize) => {
       type: DataTypes.STRING
     },
     deezer_id: {
-      type: DataTypes.INTEGER
+      type: DataTypes.BIGINT
     }
   }, {
     timestamps: true,
